Er.
A nice big--great big, really--commit here.  I guess I should commit more
often. ;)

Basically I tore apart the code and broke it up into separate files.  Once
that was done, I rejiggered the delivery code to let me write modules
to deliver to different kinds of mailboxes.  This was all because of an email
from someone asking if it would be hard to deliver to maildir mailboxes.

It's funny how just a little feature request can cause code to shatter
into lots of little pieces, isn't it?  I hope that it's a little easier to
understand now, at least.