gurgitate/mailmessage.RB: don't care about a message's encoding
Ruby 1.9 and up treat strings as having a specific encoding, but
when you're dealing with email, you won't necessarily receive a message
that's in an encoding you know (much less actually be in the encoding
it claims to have in the Content-Type: header).  So it's best just
to treat it as an arbitrary binary blob and operate on it that way.

(A MDA shouldn't care about the body of a message anyway.)